requirements_nover.txt
alabaster
amqp
anyio
appdirs
appnope
argon2-cffi
argon2-cffi-bindings
asgiref
astroid
asttokens
atomicwrites
attrs
Babel
backcall
billiard
black
boto3
botocore
cattrs
celery
certifi
cffi
cfgv
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
Collectfast
colorama
coverage
crispy-bootstrap4
decorator
defusedxml
dill
distlib
Django
django-allauth
django-anymail
django-cors-headers
django-coverage-plugin
django-crispy-forms
django-debug-toolbar
django-environ
django-extensions
django-model-utils
django-redis
django-storages
django-stubs
django-stubs-ext
djangorestframework
docutils
esbonio
exceptiongroup
executing
factory-boy
Faker
filelock
flake8
flake8-isort
gunicorn
h11
hiredis
httpie
identify
idna
imagesize
iniconfig
ipdb
ipython
ipython-genutils
isort
jedi
Jinja2
jmespath
kombu
lazy-object-proxy
livereload
lsprotocol
markdown-it-py
MarkupSafe
matplotlib-inline
mccabe
mdurl
more-itertools
multidict
mypy
mypy-extensions
nodeenv
numpy
oauthlib
packaging
parso
pathspec
pexpect
pickleshare
pillow
platformdirs
pluggy
pre-commit
prompt-toolkit
psycopg2-binary
ptyprocess
pure-eval
py
pycodestyle
pycparser
pyflakes
pygls
Pygments
pylint
pylint-django
pylint-plugin-utils
pyparsing
pyproj
PySocks
pyspellchecker
pytest
pytest-django
pytest-sugar
python-dateutil
python-slugify
python3-openid
pytz
PyYAML
redis
regex
requests
requests-oauthlib
requests-toolbelt
rich
s3transfer
six
sniffio
snowballstemmer
Sphinx
sphinx-autobuild
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sqlparse
stack-data
starlette
termcolor
testfixtures
text-unidecode
toml
tomlkit
tornado
traitlets
typed-ast
typeguard
types-PyYAML
typing_extensions
tzdata
urllib3
uvicorn
vine
virtualenv
watchfiles
wcwidth
websockets
Werkzeug
wrapt
